feat: implement crud for fiat-ramp

+++ src-tauri/migrations/5_fiat_ramp_transaction.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS fiat_ramp (
     id UUID PRIMARY KEY,
     fiat_id INTEGER NOT NULL,
-    fiat_amount DECIMAL NOT NULL,
+    fiat_amount REAL NOT NULL,
     date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     via_exchange VARCHAR(255) NOT NULL,
     kind VARCHAR(8) NOT NULL CHECK (
